# Optimize test suite, CI config