--- a/config/environments/test.rb
+++ b/config/environments/test.rb
@@ -40,4 +40,6 @@ Rails.application.configure do
   # Raises error for missing translations
   # config.action_view.raise_on_missing_translations = true
   config.action_mailer.default_url_options = { host: ENV["BASE_URL"] }
+
+  config.cache_store = :file_store, Rails.root.join("tmp", "cache", "test#{ENV["TEST_ENV_NUMBER"]}")
 end