--- a/spec/controllers/admin/bikes_controller_spec.rb
+++ b/spec/controllers/admin/bikes_controller_spec.rb
@@ -1,6 +1,6 @@
-require "spec_helper"
+require "rails_helper"
 
-describe Admin::BikesController do
+RSpec.describe Admin::BikesController do
   let(:user) { FactoryBot.create(:admin) }
   before do
     set_current_user(user)