--- a/spec/controllers/admin/dashboard_controller_spec.rb
+++ b/spec/controllers/admin/dashboard_controller_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe Admin::DashboardController do
+RSpec.describe Admin::DashboardController, type: :controller do
   describe "index" do
     context "not logged in" do
       it "redirects" do