--- a/spec/controllers/admin/failed_bikes_controller_spec.rb
+++ b/spec/controllers/admin/failed_bikes_controller_spec.rb
@@ -1,6 +1,6 @@
-require "spec_helper"
+require "rails_helper"
 
-describe Admin::FailedBikesController do
+RSpec.describe Admin::FailedBikesController do
   describe "index" do
     it "responds with OK and renders the index template" do
       user = FactoryBot.create(:admin)