--- a/spec/controllers/admin/feedbacks_controller_spec.rb
+++ b/spec/controllers/admin/feedbacks_controller_spec.rb
@@ -1,6 +1,6 @@
-require "spec_helper"
+require "rails_helper"
 
-describe Admin::FeedbacksController, type: :controller do
+RSpec.describe Admin::FeedbacksController, type: :controller do
   let(:subject) { FactoryBot.create(:feedback) }
   include_context :logged_in_as_super_admin