--- a/spec/controllers/admin/flavor_text_controller_spec.rb
+++ b/spec/controllers/admin/flavor_text_controller_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe Admin::FlavorTextsController do
+RSpec.describe Admin::FlavorTextsController, type: :controller do
   describe "destroy" do
     it "destroys" do
       text = FlavorText.create(message: "lulz")