--- a/spec/controllers/admin/mail_snippets_controller_spec.rb
+++ b/spec/controllers/admin/mail_snippets_controller_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe Admin::MailSnippetsController do
+RSpec.describe Admin::MailSnippetsController, type: :controller do
   include_context :logged_in_as_super_admin
 
   describe "index" do