--- a/spec/controllers/admin/manufacturer_controller_spec.rb
+++ b/spec/controllers/admin/manufacturer_controller_spec.rb
@@ -1,6 +1,6 @@
-require "spec_helper"
+require "rails_helper"
 
-describe Admin::ManufacturersController, type: :controller do
+RSpec.describe Admin::ManufacturersController, type: :controller do
   let(:subject) { FactoryBot.create(:manufacturer) }
   include_context :logged_in_as_super_admin