--- a/spec/controllers/admin/organization_invitations_controller_spec.rb
+++ b/spec/controllers/admin/organization_invitations_controller_spec.rb
@@ -1,4 +1,4 @@
-require "spec_helper"
+require "rails_helper"
 
 RSpec.describe Admin::OrganizationInvitationsController, type: :controller do
   let(:user) { FactoryBot.create(:admin) }