--- a/spec/controllers/admin/paid_features_controller_spec.rb
+++ b/spec/controllers/admin/paid_features_controller_spec.rb
@@ -1,6 +1,6 @@
-require "spec_helper"
+require "rails_helper"
 
-describe Admin::PaidFeaturesController, type: :controller do
+RSpec.describe Admin::PaidFeaturesController, type: :controller do
   let(:subject) { FactoryBot.create(:paid_feature) }
   include_context :logged_in_as_super_admin
   let(:passed_params) { { amount: 222.22, description: "Some really long description or wahtttt", details_link: "https://example.com", kind: "custom_one_time", name: "another name stuff" } }