--- a/spec/controllers/admin/paints_controller_spec.rb
+++ b/spec/controllers/admin/paints_controller_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe Admin::PaintsController do
+RSpec.describe Admin::PaintsController, type: :controller do
   include_context :logged_in_as_super_admin
   describe "index" do
     it "renders" do