--- a/spec/controllers/admin/recovery_displays_controller_spec.rb
+++ b/spec/controllers/admin/recovery_displays_controller_spec.rb
@@ -1,6 +1,6 @@
-require "spec_helper"
+require "rails_helper"
 
-describe Admin::RecoveryDisplaysController do
+RSpec.describe Admin::RecoveryDisplaysController do
   before do
     user = FactoryBot.create(:admin)
     set_current_user(user)