--- a/spec/controllers/admin/stolen_notifications_controller_spec.rb
+++ b/spec/controllers/admin/stolen_notifications_controller_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe Admin::StolenNotificationsController do
+RSpec.describe Admin::StolenNotificationsController, type: :controller do
   describe "index" do
     it "responds with OK and renders the index template" do
       user = FactoryBot.create(:admin)