--- a/spec/controllers/api/v1/component_types_controller_spec.rb
+++ b/spec/controllers/api/v1/component_types_controller_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe Api::V1::ComponentTypesController do
+RSpec.describe Api::V1::ComponentTypesController, type: :controller do
   describe "index" do
     it "loads the request" do
       FactoryBot.create(:ctype)