--- a/spec/controllers/api/v1/cycle_types_controller_spec.rb
+++ b/spec/controllers/api/v1/cycle_types_controller_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe Api::V1::CycleTypesController do
+RSpec.describe Api::V1::CycleTypesController, type: :controller do
   describe "index" do
     it "loads the request" do
       get :index, format: :json