--- a/spec/controllers/api/v1/handlebar_types_controller_spec.rb
+++ b/spec/controllers/api/v1/handlebar_types_controller_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe Api::V1::HandlebarTypesController do
+RSpec.describe Api::V1::HandlebarTypesController, type: :controller do
   describe "index" do
     it "loads the page" do
       get :index, format: :json