--- a/spec/controllers/api/v1/manufacturers_controller_spec.rb
+++ b/spec/controllers/api/v1/manufacturers_controller_spec.rb
@@ -1,6 +1,6 @@
-require "spec_helper"
+require "rails_helper"
 
-describe Api::V1::ManufacturersController, type: :controller do
+RSpec.describe Api::V1::ManufacturersController, type: :controller do
   describe "index" do
     it "loads the request" do
       m = FactoryBot.create(:manufacturer, name: "AAAA manufacturer")