… .circleci/config.yml
… .rspec
… Gemfile.lock
… Procfile_development
… bin/_guard-core
… bin/bundle
… bin/guard
… bin/parallel_rspec
… config/database.yml
… spec/controllers/api/v1/organizations_controller_spec.rb
… spec/controllers/api/v1/stolen_locking_response_suggestions_controller_spec.rb
… spec/controllers/api/v1/users_controller_spec.rb
… spec/controllers/api/v1/wheel_sizes_controller_spec.rb
… spec/controllers/bike_codes_controller_spec.rb
… spec/controllers/bikes/recovery_controller_spec.rb
… spec/controllers/bikes_controller_spec.rb
… spec/controllers/blogs_controller_spec.rb
… spec/controllers/discourse_authentication_controller_spec.rb
… spec/controllers/documentation_controller_spec.rb
… spec/controllers/errors_controller_spec.rb
… spec/controllers/feedbacks_controller_spec.rb
… spec/controllers/info_controller_spec.rb
… spec/controllers/integrations_controller_spec.rb
… spec/controllers/landing_pages_controller_spec.rb
… spec/controllers/locks_controller_spec.rb
… spec/controllers/manufacturers_controller_spec.rb
… spec/controllers/news_controller_spec.rb
… spec/controllers/oauth/applications_controller_spec.rb
… spec/controllers/organizations_controller_spec.rb
… spec/controllers/organized/ambassador_task_assignments_controller_spec.rb