@sethherr 2019-06-08
Do you think it would be reasonable to add some default test parallelization settings to `.env`? It seems like they have a huge benefit, so making it default to them seems reasonable.