@jmromer 2019-06-09
> Do you think it would be reasonable to add some default test parallelization settings to .env?

They won't be picked up by the shell if placed there. Luckily the env var is optional and the library's default behavior (2 processes per core) yields virtually optimal performance, and I suspect that value scales better than 1 per core as the test suite grows. I added some documentation in 10ee7ed6

> I have found render_views to be incredibly valuable, since it catches bugs in the views - and makes the controller specs closer to the integration specs I believe are the most useful.

We still render views in controller specs — that's enabled per-controller-spec-suite, in line with the library's updated default. Request specs render views.